# Conflicts:
#	cpp/subprojects/boosting/src/mlrl/boosting/losses/loss_non_decomposable_squared_error.cpp

+++ cpp/subprojects/seco/src/mlrl/seco/heuristics/heuristic_accuracy.cpp
@@ -11,12 +11,12 @@ namespace seco {
     class Accuracy final : public IHeuristic {
         public:
 
-            float64 evaluateConfusionMatrix(float64 cin, float64 cip, float64 crn, float64 crp, float64 uin,
-                                            float64 uip, float64 urn, float64 urp) const override {
-                float64 numCoveredCorrect = cin + crp;
-                float64 numUncoveredCorrect = uin + uip;
-                float64 numCorrect = numCoveredCorrect + numUncoveredCorrect;
-                float64 numTotal = numCorrect + cip + crn + urn + urp;
+            float32 evaluateConfusionMatrix(float32 cin, float32 cip, float32 crn, float32 crp, float32 uin,
+                                            float32 uip, float32 urn, float32 urp) const override {
+                float32 numCoveredCorrect = cin + crp;
+                float32 numUncoveredCorrect = uin + uip;
+                float32 numCorrect = numCoveredCorrect + numUncoveredCorrect;
+                float32 numTotal = numCorrect + cip + crn + urn + urp;
                 return util::divideOrZero(numCorrect, numTotal);
             }
     };
